chore(suite): unify staking files
