+++ packages/suite/src/components/suite/FormFractionButtons.tsx
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 import { Translation } from 'src/components/suite';
 import { Button, Tooltip } from '@trezor/components';
 import { BigNumber } from '@trezor/utils/src/bigNumber';
-import { MIN_ETH_AMOUNT_FOR_STAKING } from 'src/constants/suite/ethStaking';
+import { MIN_ETH_AMOUNT_FOR_STAKING } from '@suite-common/wallet-constants';
 import { NetworkSymbol } from '@suite-common/wallet-config';
 
 const Flex = styled.div`
