+++ suite-common/suite-utils/package.json
@@ -20,7 +20,6 @@
         "@suite-common/wallet-types": "workspace:*",
         "@trezor/connect": "workspace:*",
         "@trezor/urls": "workspace:*",
-        "date-fns": "^2.30.0",
-        "web3-utils": "^4.3.1"
+        "date-fns": "^2.30.0"
     }
 }
