+++ suite-common/suite-utils/src/index.ts
@@ -5,7 +5,6 @@ export * from './build';
 export * from './resolveStaticPath';
 export * from './comparison';
 export * from './txsPerPage';
-export * from './stake';
 export * from './parseFirmwareChangelog';
 export * from './uppercaseType';
 export * from './protocol';
