+++ suite-common/wallet-constants/package.json
@@ -9,5 +9,8 @@
         "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'",
         "depcheck": "yarn g:depcheck",
         "type-check": "yarn g:tsc --build"
+    },
+    "dependencies": {
+        "@trezor/utils": "workspace:*"
     }
 }
