+++ suite-common/wallet-constants/src/index.ts
@@ -1,3 +1,4 @@
 export * from './formDraft';
 export * from './sendForm';
 export * from './discovery';
+export * from './ethereumStakingConstants';
