+++ suite-common/wallet-core/src/index.ts
@@ -37,3 +37,4 @@ export * from './stake/stakeSelectors';
 export * from './stake/stakeMiddleware';
 export * from './stake/stakeThunks';
 export * from './stake/stakeTypes';
+export * from './stake/stakeConstants';
