+++ suite-common/wallet-core/src/transactions/transactionsReducer.ts
@@ -7,12 +7,15 @@ import {
     getConfirmations,
     isPending,
     getEverstakePool,
+    isClaimTx,
+    isStakeTx,
+    isStakeTypeTx,
+    isUnstakeTx,
 } from '@suite-common/wallet-utils';
 import {
     getIsPhishingTransaction,
     TokenDefinitionsRootState,
 } from '@suite-common/token-definitions';
-import { isClaimTx, isStakeTx, isStakeTypeTx, isUnstakeTx } from '@suite-common/suite-utils';
 import { createReducerWithExtraDeps } from '@suite-common/redux-utils';
 import { selectNetworkTokenDefinitions } from '@suite-common/token-definitions/src/tokenDefinitionsSelectors';
