+++ suite-common/wallet-types/src/ethereumStaking.ts
@@ -1,0 +1,5 @@
+export type StakeType = 'stake' | 'unstake' | 'claim';
+
+export const supportedNetworkSymbols = ['eth', 'thol'] as const;
+
+export type SupportedNetworkSymbol = (typeof supportedNetworkSymbols)[number];
