--- suite-common/wallet-types/src/stake.ts
@@ -1,1 +1,0 @@
-export type StakeType = 'stake' | 'unstake' | 'claim';
